+++ scripts/travis.sh
@@ -33,7 +33,7 @@ run ./test/mock/defaults-change.sh
 run ./test/mock/short-versions.sh
 run ./test/mock/direct-paths.sh
 run ./test/mock/expected_ports.sh
-run ./test/mock/read-only-replication.sh
+run ./test/mock/replication-setup.sh
 run ./test/mock/ndb_test.sh
 run ./test/mock/pxc_test.sh
 run ./test/mock/cookbook.sh